refactor: move response_content into backend code
