+++ gitlab/v4/objects/files.py
@@ -272,7 +272,7 @@ class ProjectFileManager(GetMixin, CreateMixin, UpdateMixin, DeleteMixin, RESTMa
         )
         if TYPE_CHECKING:
             assert isinstance(result, requests.Response)
-        return utils.response_content(
+        return self.gitlab._backend.response_content(
             result, streamed, action, chunk_size, iterator=iterator
         )
